Revert to monolithic UI app structure for Docker compatibility
