--- ui/api/__init__.py
@@ -1,1 +1,0 @@
-# API client package
